forage performance improvements
