+++ package.json
@@ -4,13 +4,14 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+    "build": "next build && yarn initNames",
     "test": "vitest run --silent=false",
     "test:pure": "yarn vitest run ./utils/test/level.test.ts --silent=false",
     "test:server": "yarn vitest run ./utils/test/server.test.ts --silent=false",
     "start": "next start",
     "lint": "next lint",
-    "initDB": "npx ts-node --skip-project ./server/database/migrations/migration-1.ts"
+    "initDB": "npx ts-node --skip-project ./server/database/migrations/migration-1.ts",
+    "initNames": "npx ts-node --skip-project ./scripts/init.ts"
   },
   "dependencies": {
     "@auth0/nextjs-auth0": "^1.9.2",
