+++ server/fileSystemData/mushroomNames.json
@@ -1,0 +1,1 @@
+{"mushroomNames":["agaricus-macrosporus","bay-bolete","blusher","blushing-wood-mushroom","chestnut-bolete","field","grey-spotted-amanita","honey-fungus","horse","hortiboletus-bubalinus","jersey-cow-fungus","larch-bolete","medusa","oak-bolete","orange-bolete","orange-grisette","pavement","peppery-bolete","prince","red-cracked-bolete","scarletina-bolete","snakeskin-grisette ","spring-fieldcap","tawny-grisette","the-great-wood-mushroom","wood-mushroom"]}
